+++ tests/conftest.py
@@ -11,6 +11,7 @@ from functools import partial
 import coverage
 import pytest
 import six
+from virtualenv.discovery.builtin import get_interpreter
 
 from virtualenv.discovery.py_info import PythonInfo
 from virtualenv.info import IS_PYPY, IS_WIN, fs_supports_symlink
@@ -318,3 +319,15 @@ def temp_app_data(monkeypatch, tmp_path):
     app_data = tmp_path / "app-data"
     monkeypatch.setenv(str("VIRTUALENV_OVERRIDE_APP_DATA"), str(app_data))
     return app_data
+
+
+@pytest.fixture(scope="session")
+def cross_python(is_inside_ci, session_app_data):
+    spec = "{}{}".format(sys.implementation, 2 if sys.version_info[0] == 3 else 3)
+    interpreter = get_interpreter(spec, session_app_data)
+    if interpreter is None:
+        msg = "could not find {}".format(spec)
+        if is_inside_ci:
+            raise RuntimeError(msg)
+        pytest.skip(msg=msg)
+    yield interpreter
